requirement.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
sqlparse==0.5.3
tzdata==2024.2
XlsxWriter==3.2.3
